created new file map_builder.c ad 100% norm fixed in reader.c
