+++ src/parser/check_1.c
@@ -6,7 +6,7 @@
 /*   By: framos-p <framos-p@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/30 11:55:26 by framos-p          #+#    #+#             */
-/*   Updated: 2023/11/30 12:07:02 by framos-p         ###   ########.fr       */
+/*   Updated: 2023/12/04 11:19:11 by framos-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
